# Fix bug in per-segment rows calculation when a child is replicated