--- a/src/backend/gporca/data/dxl/minidump/ReplicatedLOJHashDistributedTable.mdp
+++ b/src/backend/gporca/data/dxl/minidump/ReplicatedLOJHashDistributedTable.mdp
@@ -1420,7 +1420,7 @@
     <dxl:Plan Id="0" SpaceSize="6">
       <dxl:GatherMotion InputSegments="0,1,2" OutputSegments="-1">
         <dxl:Properties>
-          <dxl:Cost StartupCost="0" TotalCost="862.030854" Rows="100.000001" Width="16"/>
+          <dxl:Cost StartupCost="0" TotalCost="862.058851" Rows="100.000001" Width="16"/>
         </dxl:Properties>
         <dxl:ProjList>
           <dxl:ProjElem ColId="0" Alias="a">
@@ -1440,7 +1440,7 @@
         <dxl:SortingColumnList/>
         <dxl:HashJoin JoinType="Left">
           <dxl:Properties>
-            <dxl:Cost StartupCost="0" TotalCost="862.024892" Rows="100.000001" Width="16"/>
+            <dxl:Cost StartupCost="0" TotalCost="862.040963" Rows="100.000001" Width="16"/>
           </dxl:Properties>
           <dxl:ProjList>
             <dxl:ProjElem ColId="0" Alias="a">
@@ -1466,7 +1466,7 @@
           </dxl:HashCondList>
           <dxl:RedistributeMotion InputSegments="0" OutputSegments="0,1,2" DuplicateSensitive="true">
             <dxl:Properties>
-              <dxl:Cost StartupCost="0" TotalCost="431.005655" Rows="100.000000" Width="8"/>
+              <dxl:Cost StartupCost="0" TotalCost="431.006082" Rows="100.000000" Width="8"/>
             </dxl:Properties>
             <dxl:ProjList>
               <dxl:ProjElem ColId="0" Alias="a">