… src/backend/gporca/data/dxl/minidump/CTE15Replicated.mdp
… src/backend/gporca/data/dxl/minidump/CTE2HAReplicated.mdp
… src/backend/gporca/data/dxl/minidump/CTE2Replicated.mdp
… src/backend/gporca/data/dxl/minidump/FullJoin-Replicated.mdp
… src/backend/gporca/data/dxl/minidump/NonSingleton.mdp
… src/backend/gporca/data/dxl/minidump/ReplicatedJoinHashDistributedTable.mdp
… src/backend/gporca/data/dxl/minidump/ReplicatedJoinRandomDistributedTable.mdp
… src/backend/gporca/data/dxl/minidump/ReplicatedTableCTE.mdp
… src/test/regress/expected/bfv_joins.out
… src/test/regress/expected/bfv_joins_optimizer.out
… src/test/regress/sql/bfv_joins.sql