Orca's behavior in setting the operator's per-segment rows is incorrect when the child is replicated. As a result, the hash join often chooses the [hash, hash]alternative, even when better alternatives are available.

Example:
create table rep as select generate_series(1,10000) a distributed replicated;
create table rand as select generate_series(1,100) a distributed randomly;
analyze rep, rand;
explain select * from rand t1 join rep t2 on t1.a = t2.a; PhysicalPlan:
```
+--CPhysicalMotionGather(coordinator)
   +--CPhysicalInnerHashJoin
      |--CPhysicalMotionHashDistribute HASHED
      |  +--CPhysicalTableScan "rep" ("rep")
      |--CPhysicalMotionHashDistribute HASHED
      |  +--CPhysicalTableScan "rand" ("rand")
      +--CScalarCmp (=)
         |--CScalarIdent "a" (0)
         +--CScalarIdent "a" (8)
```
Note: Number of segments=3
In the above example, with regards to the operator CPhysicalMotionHashDistribute over the 'rep' table, we are incorrectly setting  the per-segment number of rows to 3333 instead of 10000. This incorrect setting is the reason for the lower cost of the hash join alternative (redistribute, redistribute). This PR addresses the issue of adjusting per-segment row calculations when the operator's child is replicated. After applying the fix, it was noticed that the query's execution time decreased compared to the previous run without fix.

Before fix:

explain analyze select * from rand t1 join rep t2 on t1.a = t2.a;
```

                                                                    QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)  (cost=0.00..862.96 rows=100 width=8) (actual time=18.842..19.111 rows=100 loops=1)
   ->  Hash Join  (cost=0.00..862.96 rows=34 width=8) (actual time=9.693..14.288 rows=38 loops=1)
         Hash Cond: (rep.a = rand.a)
         Extra Text: (seg0)   Hash chain length 1.0 avg, 1 max, using 38 of 524288 buckets.
         ->  Result  (cost=0.00..431.37 rows=3334 width=4) (actual time=0.180..2.964 rows=3385 loops=1)
               ->  Seq Scan on rep  (cost=0.00..431.19 rows=10000 width=4) (actual time=0.168..1.077 rows=10000 loops=1)
         ->  Hash  (cost=431.00..431.00 rows=34 width=4) (actual time=2.770..2.770 rows=38 loops=1)
               Buckets: 524288  Batches: 1  Memory Usage: 4098kB
               ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=0.00..431.00 rows=34 width=4) (actual time=0.950..2.744 rows=38 loops=1)
                     Hash Key: rand.a
                     ->  Seq Scan on rand  (cost=0.00..431.00 rows=34 width=4) (actual time=0.157..0.161 rows=37 loops=1)
 Optimizer: GPORCA
 Planning Time: 11.404 ms
 Execution Time: 95.326 ms
```

 After Fix:
explain analyze select * from rand t1 join rep t2 on t1.a = t2.a;
```

                                                                   QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------------
 Gather Motion 1:1  (slice1; segments: 1)  (cost=0.00..864.06 rows=100 width=8) (actual time=9.607..9.627 rows=100 loops=1)
   ->  Hash Join  (cost=0.00..864.05 rows=300 width=8) (actual time=3.345..9.015 rows=100 loops=1)
         Hash Cond: (rep.a = rand.a)
         Extra Text: Hash chain length 1.0 avg, 1 max, using 100 of 524288 buckets.
         ->  Seq Scan on rep  (cost=0.00..431.19 rows=30000 width=4) (actual time=0.226..1.846 rows=10000 loops=1)
         ->  Hash  (cost=431.01..431.01 rows=300 width=4) (actual time=0.271..0.272 rows=100 loops=1)
               Buckets: 524288  Batches: 1  Memory Usage: 4100kB
               ->  Broadcast Motion 3:1  (slice2; segments: 3)  (cost=0.00..431.01 rows=300 width=4) (actual time=0.023..0.227 rows=100 loops=1)
                     ->  Seq Scan on rand  (cost=0.00..431.00 rows=34 width=4) (actual time=0.884..0.891 rows=40 loops=1)
 Optimizer: GPORCA
 Planning Time: 16.417 ms
 Execution Time: 31.988 ms
```